+++ noodles-cram/src/crai/async/fs.rs
@@ -4,8 +4,8 @@ use std::path::Path;
 
 use tokio::{fs::File, io};
 
-use super::io::Reader;
-use crate::crai::Index;
+use super::io::{Reader, Writer};
+use crate::crai::{Index, Record};
 
 /// Reads the entire contents of a CRAM index.
 ///
@@ -28,4 +28,30 @@ where
 {
     let mut reader = File::open(src).await.map(Reader::new)?;
     reader.read_index().await
+}
+
+/// Writes a CRAM index to a file.
+///
+/// This is a convenience function and is equivalent to creating a file at the given path and
+/// writing the index.
+///
+/// # Examples
+///
+/// ```no_run
+/// # #[tokio::main]
+/// # async fn main() -> tokio::io::Result<()> {
+/// use noodles_cram::crai;
+/// let index = crai::Index::default();
+/// crai::r#async::fs::write("sample.cram.crai", &index).await?;
+/// # Ok(())
+/// # }
+/// ```
+pub async fn write<P>(dst: P, index: &[Record]) -> tokio::io::Result<()>
+where
+    P: AsRef<Path>,
+{
+    let mut writer = File::create(dst).await.map(Writer::new)?;
+    writer.write_index(index).await?;
+    writer.shutdown().await?;
+    Ok(())
 }